Clean up unused contexts
