--- apps/smart-forms-app/src/features/debug/contexts/DebugModeContext.tsx
@@ -1,47 +1,0 @@
-/*
- * Copyright 2023 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { ReactNode } from 'react';
-import { createContext, useState } from 'react';
-import { SHOW_DEBUG_MODE } from '../../../utils/env.ts';
-
-export type DebugModeContextType = {
-  isDebugMode: boolean;
-  activateDebugMode: () => unknown;
-};
-
-export const DebugModeContext = createContext<DebugModeContextType>({
-  isDebugMode: false,
-  activateDebugMode: () => void 0
-});
-
-const debugModeEnv: string = SHOW_DEBUG_MODE;
-
-function DebugModeContextProvider(props: { children: ReactNode }) {
-  const { children } = props;
-
-  const [isDebugMode, setIsDebugMode] = useState(debugModeEnv === 'true');
-
-  const debugModeContext: DebugModeContextType = {
-    isDebugMode: isDebugMode,
-    activateDebugMode: () => setIsDebugMode(true)
-  };
-
-  return <DebugModeContext.Provider value={debugModeContext}>{children}</DebugModeContext.Provider>;
-}
-
-export default DebugModeContextProvider;
