+++ apps/smart-forms-app/src/features/preprocess/utils/preprocessQuestionnaire/resolveValueSets.ts
@@ -21,9 +21,9 @@ import {
   getValueSetCodings,
   getValueSetPromise,
   resolvePromises
-} from '../../../valueSet/utils/valueSet.ts';
+} from '../../../../utils/valueSet.ts';
 import type { Variables } from '../../../../providers/questionnaireProvider.interfaces.ts';
-import type { ValueSetPromise } from '../../../valueSet/types/valueSet.interface.ts';
+import type { ValueSetPromise } from '../../../../types/valueSet.interface.ts';
 
 export async function resolveValueSets(
   variables: Variables,
