--- apps/smart-forms-app/src/features/smartAppLaunch/contexts/SmartAppLaunchContext.tsx
@@ -1,68 +1,0 @@
-/*
- * Copyright 2023 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { ReactNode } from 'react';
-import { createContext, useState } from 'react';
-import type Client from 'fhirclient/lib/Client';
-import type { Encounter, Patient, Practitioner } from 'fhir/r4';
-import type { SmartAppLaunchContextType } from '../types/smartAppLaunch.type.ts';
-
-export const SmartAppLaunchContext = createContext<SmartAppLaunchContextType>({
-  fhirClient: null,
-  patient: null,
-  user: null,
-  encounter: null,
-  setFhirClient: () => void 0,
-  setPatient: () => void 0,
-  setUser: () => void 0,
-  setEncounter: () => void 0
-});
-
-function SmartAppLaunchContextProvider(props: { children: ReactNode }) {
-  const { children } = props;
-  const [client, setClient] = useState<Client | null>(null);
-  const [patient, setPatient] = useState<Patient | null>(null);
-  const [user, setUser] = useState<Practitioner | null>(null);
-  const [encounter, setEncounter] = useState<Encounter | null>(null);
-
-  const launchContext: SmartAppLaunchContextType = {
-    fhirClient: client,
-    patient: patient,
-    user: user,
-    encounter: encounter,
-    setFhirClient: (client: Client) => {
-      setClient(client);
-    },
-    setPatient: (patient: Patient) => {
-      setPatient(patient);
-    },
-    setUser: (user: Practitioner) => {
-      setUser(user);
-    },
-    setEncounter: (encounter: Encounter) => {
-      setEncounter(encounter);
-    }
-  };
-
-  return (
-    <SmartAppLaunchContext.Provider value={launchContext}>
-      {children}
-    </SmartAppLaunchContext.Provider>
-  );
-}
-
-export default SmartAppLaunchContextProvider;
