--- apps/smart-forms-app/src/features/smartAppLaunch/types/smartAppLaunch.type.ts
@@ -1,30 +1,0 @@
-/*
- * Copyright 2023 Commonwealth Scientific and Industrial Research
- * Organisation (CSIRO) ABN 41 687 119 230.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import type { Encounter, Patient, Practitioner } from 'fhir/r4';
-import type Client from 'fhirclient/lib/Client';
-
-export type SmartAppLaunchContextType = {
-  fhirClient: Client | null;
-  patient: Patient | null;
-  user: Practitioner | null;
-  encounter: Encounter | null;
-  setFhirClient: (client: Client) => unknown;
-  setPatient: (patient: Patient) => unknown;
-  setUser: (user: Practitioner) => unknown;
-  setEncounter: (user: Encounter) => unknown;
-};
